refactor(Twitter): Arrange packages
